requirements.txt
gunicorn
python-dotenv
Flask>2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
requests==2.27.1
binance-futures-connector==4.1.0